Consolidate and document SQL AST shims
